requirements.txt
Django==3.2.19
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
python-decouple==3.8
django-redis